requirements.txt
asgiref==3.8.1
asttokens==3.0.0
backcall==0.2.0
backports.zoneinfo==0.2.1
coverage==7.6.1
decorator==5.1.1
Django==4.2.18
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.2.0
ipdb==0.13.13
ipython==8.12.3
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
PyJWT==2.9.0
sqlparse==0.5.3
stack-data==0.6.3
tomli==2.2.1
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
